Actually call the ready() function.
Count the number of paths we are listening to.
Add the file listener to the communicator only if it has a valid socket.
Check that the secure-listen/certifcate/private-key are defined before getting the string.
If the snaprfs does not listen to any file, then the start fails.
Added configuration logs to ease the potential problems in the paths configuration files.
Properly name the file_listener connection.