Handle remote connection correctly.
The remote snaprfs now receives one or two IPs: plain & secure <ip:port>
addresses. If we received a message through a secure communicator daemon
then we must use the "secure" <ip:port> address to connect to the snaprfs.
This version does that (still need to properly detect the secure connection
in the communicator).

Added some names for parameter names.