Cleaned up the dat a receiver.
The receiver would attempt to read some data from the header, then from the
footer once. Now there is a loop and if we get no data before completion,
we return early.

Now each block returns early if the load was not yet complete so there is
no need to do that in the following block.

Fixed the server searching for a file, I had to check == end and not != end.