Renamed the "--parts" parameter.
The rsync tool calls that the tempdir, so I switched to --temp-dirs instead
if --parts. I think that will be much clearer to everyone. I also updated
the documentation of that parameter in the snaprfs.conf. Again, I think that
will be clearer.

Fixed the reference to TODO.md (was still named TODO.txt in debian/docs).