Added support for files to be modified and still opened.
Now, if you open a file, make changes to it and keep it opened. The snaprfs
will still react to the change after a little while (10 seconds by default)
and transfer the file. This is useful for files such as logs.

The type of file is also checked. If not regular, directory or symbolic link,
then it does not get transferred at all.